Listeners should register a queue matching their own class name.
When you are using multiple listeners or subscribers listening to the same model,
both queues would have the same name, meaning that messages would be pushed to
each listener randomly.

This commit modifies Listener so that it registers a queue with it's own class
name rather than the class it's listening to.