outbox: reset publishing channel
channel publish performance appears to degrade over time, resetting
the channel appears to fix the problem. On a sample app I measured
a decrease from ~300/s to ~220/s over an hour period (roughly the same
rate for longer periods of time)

I do not know what the root cause is (i.e. is this an issue with RMQ,
the ruby bunny library or something application level) resetting the
channel seems to fix it (keep publishing rate near max). This operation
is very cheap and quick so it is probably easier to reset every few minutes until
there is more time to investigate the root cause.